requirements.txt
torch>=1.7.0
torchvision>=0.8.1
numpy>=1.19.0
Pillow>=8.0.0
opencv-python>=4.4.0
transformers>=4.0.0
u2net-pytorch>=0.0.1
requests